Share variable width operator implementations where possible
